refactor(cli): flatten commands, delete global workspace model
Delete workspace-command.ts (global workspace management: symlinks, jsrepo
install to ~/.epicenter/workspaces/) and data-command.ts (nested data
subcommand group).

Extract each handler into individual top-level command files:
get, list, count, delete, tables, kv, export.

Add shared helpers in util/with-workspace-options.ts:
withWorkspaceOptions (adds --dir, --workspace, --format),
runDataCommand (open/use/dispose lifecycle), resolveTable.

Remove workspacesDir export from index.ts and util/paths.ts.
