+++ packages/cli/src/util/paths.ts
@@ -5,7 +5,3 @@ import { join } from 'node:path';
 export function resolveEpicenterHome(flagValue?: string): string {
 	return flagValue ?? Bun.env.EPICENTER_HOME ?? join(homedir(), '.epicenter');
 }
-
-export function workspacesDir(home: string): string {
-	return join(home, 'workspaces');
-}
